# stringify only when necessary