Benchmark setting span and transaction context.
Before
```
$ go test -bench=.
goos: linux
goarch: amd64
pkg: go.elastic.co/apm/module/apmot
BenchmarkSetSpanContext-4          	  300000	      4201 ns/op
BenchmarkSetTransactionContext-4   	  300000	      4593 ns/op
PASS
ok  	go.elastic.co/apm/module/apmot	2.881s
```
After
```
$ go test -bench=.
goos: linux
goarch: amd64
pkg: go.elastic.co/apm/module/apmot
BenchmarkSetSpanContext-4          	 1000000	      2148 ns/op
BenchmarkSetTransactionContext-4   	  500000	      2297 ns/op
PASS
ok  	go.elastic.co/apm/module/apmot	3.803s
```